refactor activity dates code
